Chatbackend/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
openai==1.75.0
Werkzeug==3.0.1 
schedule==1.2.2
Flask-PyMongo==2.3.0
numpy
Flask-SocketIO==5.3.6
yt-dlp
pytube
pysrt
gunicorn
celery[redis]>=5.0
